[INLONG-7004][Manager] Fix UT error
